requirements.txt
gradio
torch
transformers
Pillow
requests
